smtp::Envelope.to_addresses
This *should* be just the addresses for the current MX server, but we still need to
do a major refactor to make that happen properly.

Previous code in smtp pulled out all of the To addresses from the Email. But these are
too many, and the parsing is redundant since our list of recipients in InternalStatus
already has these.

worker/mod still needs to deliver per MX server for code to work again.